+++ src/main/java/com/etiya/academy/dto/product/GetProductByIdResponseDto.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class GetProductByIdResponseDto {
     private String name;
-    private double unitPrice;
+    private BigDecimal unitPrice;
     private int unitsInStock;
     private int categoryId;
 }
